refactor: rename hook
